Fix warnings and error for run lint
